split snapshot loader
