+++ frontend/src/scenes/session-recordings/player/snapshotDataLogic.tsx
@@ -1,0 +1,341 @@
+import { actions, beforeUnmount, connect, defaults, kea, key, listeners, path, props, reducers, selectors } from 'kea'
+import { loaders } from 'kea-loaders'
+import posthog from 'posthog-js'
+
+import '@posthog/rrweb-types'
+
+import api from 'lib/api'
+import { FEATURE_FLAGS } from 'lib/constants'
+import 'lib/dayjs'
+import { FeatureFlagsSet, featureFlagLogic } from 'lib/logic/featureFlagLogic'
+import { parseEncodedSnapshots } from 'scenes/session-recordings/player/snapshot-processing/process-all-snapshots'
+import { keyForSource } from 'scenes/session-recordings/player/snapshot-processing/source-key'
+
+import '~/queries/utils'
+import {
+    RecordingSnapshot,
+    SessionRecordingId,
+    SessionRecordingSnapshotParams,
+    SessionRecordingSnapshotSource,
+    SessionRecordingSnapshotSourceResponse,
+    SessionRecordingType,
+    SnapshotSourceType,
+} from '~/types'
+
+import type { snapshotDataLogicType } from './snapshotDataLogicType'
+
+// +- before and after start and end of a recording to query for session linked events.
+// +- before and after start and end of a recording to query for events related by person.
+const DEFAULT_REALTIME_POLLING_MILLIS = 3000
+const DEFAULT_V2_POLLING_INTERVAL_MS = 10000
+
+export interface SnapshotLogicProps {
+    sessionRecordingId: SessionRecordingId
+    // allows altering v1 polling interval in tests
+    realTimePollingIntervalMilliseconds?: number
+    // allows disabling polling for new sources in tests
+    blobV2PollingDisabled?: boolean
+    playerKey?: string
+    accessToken?: string
+}
+
+export const snapshotDataLogic = kea<snapshotDataLogicType>([
+    path((key) => ['scenes', 'session-recordings', 'snapshotLogic', key]),
+    props({} as SnapshotLogicProps),
+    key(({ sessionRecordingId }) => sessionRecordingId || 'no-session-recording-id'),
+    connect(() => ({
+        values: [featureFlagLogic, ['featureFlags']],
+    })),
+    defaults({
+        sessionPlayerMetaData: null as SessionRecordingType | null,
+    }),
+    actions({
+        loadSnapshots: true,
+        loadSnapshotSources: (breakpointLength?: number) => ({ breakpointLength }),
+        loadNextSnapshotSource: true,
+        loadSnapshotsForSource: (sources: Pick<SessionRecordingSnapshotSource, 'source' | 'blob_key'>[]) => ({
+            sources,
+        }),
+        pollRealtimeSnapshots: true,
+        stopRealtimePolling: true,
+    }),
+    reducers(() => ({
+        isRealtimePolling: [
+            false as boolean,
+            {
+                pollRealtimeSnapshots: () => true,
+                stopRealtimePolling: () => false,
+            },
+        ],
+
+        snapshotsBySourceSuccessCount: [
+            0,
+            {
+                loadSnapshotsForSourceSuccess: (state) => state + 1,
+            },
+        ],
+    })),
+    loaders(({ values, props, cache }) => ({
+        snapshotSources: [
+            null as SessionRecordingSnapshotSource[] | null,
+            {
+                loadSnapshotSources: async ({ breakpointLength }, breakpoint) => {
+                    if (breakpointLength) {
+                        await breakpoint(breakpointLength)
+                    }
+
+                    const headers: Record<string, string> = {}
+                    if (props.accessToken) {
+                        headers.Authorization = `Bearer ${props.accessToken}`
+                    }
+
+                    const blob_v2 =
+                        values.featureFlags[FEATURE_FLAGS.RECORDINGS_BLOBBY_V2_REPLAY] || !!props.accessToken
+                    const blob_v2_lts =
+                        values.featureFlags[FEATURE_FLAGS.RECORDINGS_BLOBBY_V2_LTS_REPLAY] || !!props.accessToken
+                    const response = await api.recordings.listSnapshotSources(
+                        props.sessionRecordingId,
+                        {
+                            blob_v2,
+                            blob_v2_lts,
+                        },
+                        headers
+                    )
+
+                    if (!response.sources) {
+                        return []
+                    }
+                    const anyBlobV2 = response.sources.some((s) => s.source === SnapshotSourceType.blob_v2)
+
+                    if (anyBlobV2) {
+                        return response.sources.filter((s) => s.source === SnapshotSourceType.blob_v2)
+                    }
+                    return response.sources.filter((s) => s.source !== SnapshotSourceType.blob_v2)
+                },
+            },
+        ],
+        snapshotsForSource: [
+            null as SessionRecordingSnapshotSourceResponse | null,
+            {
+                loadSnapshotsForSource: async ({ sources }, breakpoint) => {
+                    let params: SessionRecordingSnapshotParams
+
+                    if (sources.length > 1) {
+                        // they all have to be blob_v2
+                        if (sources.some((s) => s.source !== SnapshotSourceType.blob_v2)) {
+                            throw new Error('Unsupported source for multiple sources')
+                        }
+                        params = {
+                            source: 'blob_v2',
+                            // so the caller has to make sure these are in order!
+                            start_blob_key: sources[0].blob_key,
+                            end_blob_key: sources[sources.length - 1].blob_key,
+                        }
+                    } else {
+                        const source = sources[0]
+
+                        if (source.source === SnapshotSourceType.blob) {
+                            if (!source.blob_key) {
+                                throw new Error('Missing key')
+                            }
+                            params = { blob_key: source.blob_key, source: 'blob' }
+                        } else if (source.source === SnapshotSourceType.realtime) {
+                            params = { source: 'realtime' }
+                        } else if (source.source === SnapshotSourceType.blob_v2) {
+                            params = { source: 'blob_v2', blob_key: source.blob_key }
+                        } else if (source.source === SnapshotSourceType.file) {
+                            // no need to load a file source, it is already loaded
+                            return { source }
+                        } else {
+                            throw new Error(`Unsupported source: ${source.source}`)
+                        }
+                    }
+
+                    await breakpoint(1)
+
+                    const headers: Record<string, string> = {}
+                    if (props.accessToken) {
+                        headers.Authorization = `Bearer ${props.accessToken}`
+                    }
+                    const response = await api.recordings
+                        .getSnapshots(props.sessionRecordingId, params, headers)
+                        .catch((e) => {
+                            if (sources[0].source === 'realtime' && e.status === 404) {
+                                // Realtime source is not always available, so a 404 is expected
+                                return []
+                            }
+                            throw e
+                        })
+
+                    // sorting is very cheap for already sorted lists
+                    const parsedSnapshots = (await parseEncodedSnapshots(response, props.sessionRecordingId)).sort(
+                        (a, b) => a.timestamp - b.timestamp
+                    )
+                    // we store the data in the cache because we want to avoid copying this data as much as possible
+                    // and kea's immutability means we were copying all the data on every snapshot call
+                    cache.snapshotsBySource = cache.snapshotsBySource || {}
+                    // it doesn't matter which source we use as the key, since we combine the snapshots anyway
+                    cache.snapshotsBySource[keyForSource(sources[0])] = { snapshots: parsedSnapshots }
+                    // but we do want to mark the sources as loaded
+                    sources.forEach((s) => {
+                        const k = keyForSource(s)
+                        // we just need something against each key so we don't load it again
+                        cache.snapshotsBySource[k] = cache.snapshotsBySource[k] || {}
+                        cache.snapshotsBySource[k].sourceLoaded = true
+                    })
+
+                    return { sources: sources }
+                },
+            },
+        ],
+    })),
+    listeners(({ values, actions, cache, props }) => ({
+        loadSnapshots: () => {
+            // This kicks off the loading chain
+            if (!values.snapshotSourcesLoading) {
+                actions.loadSnapshotSources()
+            }
+        },
+
+        loadSnapshotSourcesSuccess: () => {
+            // When we receive the list of sources, we can kick off the loading chain
+            actions.loadNextSnapshotSource()
+        },
+
+        loadSnapshotsForSourceSuccess: ({ snapshotsForSource }) => {
+            const sources = values.snapshotSources
+            const sourceKey = snapshotsForSource.sources
+                ? keyForSource(snapshotsForSource.sources[0])
+                : keyForSource(snapshotsForSource.source)
+            const snapshots = (cache.snapshotsBySource || {})[sourceKey] || []
+
+            // Cache the last response count to detect if we're getting the same data over and over
+            const newSnapshotsCount = snapshots.length
+
+            if ((cache.lastSnapshotsCount ?? newSnapshotsCount) === newSnapshotsCount) {
+                // if we're getting no results from realtime polling, we can increment faster
+                // so that we stop polling sooner
+                const increment = newSnapshotsCount === 0 ? 2 : 1
+                cache.lastSnapshotsUnchangedCount = (cache.lastSnapshotsUnchangedCount ?? 0) + increment
+            } else {
+                cache.lastSnapshotsUnchangedCount = 0
+            }
+            cache.lastSnapshotsCount = newSnapshotsCount
+
+            if (!snapshots.length && sources?.length === 1 && sources[0].source !== SnapshotSourceType.file) {
+                // We got only a single source to load, loaded it successfully, but it had no snapshots.
+                posthog.capture('recording_snapshots_v2_empty_response', {
+                    source: sources[0],
+                })
+            }
+
+            actions.loadNextSnapshotSource()
+        },
+
+        loadNextSnapshotSource: () => {
+            // yes this is ugly duplication, but we're going to deprecate v1 and I want it to be clear which is which
+            if (values.snapshotSources?.some((s) => s.source === SnapshotSourceType.blob_v2)) {
+                const nextSourcesToLoad =
+                    values.snapshotSources?.filter((s) => {
+                        const sourceKey = keyForSource(s)
+                        return (
+                            !cache.snapshotsBySource?.[sourceKey]?.sourceLoaded && s.source !== SnapshotSourceType.file
+                        )
+                    }) || []
+
+                if (nextSourcesToLoad.length > 0) {
+                    return actions.loadSnapshotsForSource(nextSourcesToLoad.slice(0, 30))
+                }
+
+                if (!props.blobV2PollingDisabled) {
+                    actions.loadSnapshotSources(DEFAULT_V2_POLLING_INTERVAL_MS)
+                }
+            } else {
+                const nextSourceToLoad = values.snapshotSources?.find((s) => {
+                    const sourceKey = keyForSource(s)
+                    return !cache.snapshotsBySource?.[sourceKey]?.sourceLoaded && s.source !== SnapshotSourceType.file
+                })
+
+                if (nextSourceToLoad) {
+                    return actions.loadSnapshotsForSource([nextSourceToLoad])
+                }
+
+                // If we have a realtime source, start polling it
+                const realTimeSource = values.snapshotSources?.find((s) => s.source === SnapshotSourceType.realtime)
+                if (realTimeSource) {
+                    actions.pollRealtimeSnapshots()
+                }
+            }
+
+            // actions.reportUsageIfFullyLoaded()
+        },
+        pollRealtimeSnapshots: () => {
+            // always make sure we've cleared up the last timeout
+            clearTimeout(cache.realTimePollingTimeoutID)
+            cache.realTimePollingTimeoutID = null
+
+            // ten is an arbitrary limit to try to avoid sending requests to our backend unnecessarily
+            // we could change this or add to it e.g. only poll if browser is visible to user
+            if ((cache.lastSnapshotsUnchangedCount ?? 0) <= 10) {
+                cache.realTimePollingTimeoutID = setTimeout(() => {
+                    actions.loadSnapshotsForSource([{ source: SnapshotSourceType.realtime }])
+                }, props.realTimePollingIntervalMilliseconds || DEFAULT_REALTIME_POLLING_MILLIS)
+            } else {
+                actions.stopRealtimePolling()
+            }
+        },
+    })),
+    selectors(({ cache }) => ({
+        snapshotsLoading: [
+            (s) => [s.snapshotSourcesLoading, s.snapshotsForSourceLoading, s.featureFlags, s.snapshotsBySources],
+            (
+                snapshotSourcesLoading: boolean,
+                snapshotsForSourceLoading: boolean,
+                featureFlags: FeatureFlagsSet,
+                snapshotsBySources: Record<string, RecordingSnapshot[]>
+            ): boolean => {
+                const snapshots = Object.values(snapshotsBySources).flat()
+                // For v2 recordings, only show loading if we have no snapshots AND we're actually loading something.
+                if (featureFlags[FEATURE_FLAGS.RECORDINGS_BLOBBY_V2_REPLAY]) {
+                    return snapshots?.length === 0 && (snapshotSourcesLoading || snapshotsForSourceLoading)
+                }
+
+                // Default behavior for non-v2 recordings
+                // if there's a realTimePollingTimeoutID, don't signal that we're loading
+                // we don't want the UI to flip to "loading" every time we poll
+                return !cache.realTimePollingTimeoutID && (snapshotSourcesLoading || snapshotsForSourceLoading)
+            },
+        ],
+
+        snapshotsLoaded: [(s) => [s.snapshotSources], (snapshotSources): boolean => !!snapshotSources],
+
+        // rawSnapshotsFullyLoaded: [
+        //     (s) => [s.rawSnapshots, s.sessionPlayerMetaDataLoading, s.snapshotsLoading],
+        //     (rawSnapshots, sessionPlayerMetaDataLoading, snapshotsLoading): boolean => {
+        //         // TODO: Do a proper check for all sources having been loaded
+        //         return !!rawSnapshots?.length && !sessionPlayerMetaDataLoading && !snapshotsLoading
+        //     },
+        // ],
+
+        snapshotsBySources: [
+            (s) => [s.snapshotsBySourceSuccessCount],
+            (
+                // oxlint-disable-next-line @typescript-eslint/no-unused-vars
+                _snapshotsBySourceSuccessCount: number
+            ): Record<string, RecordingSnapshot[]> => {
+                if (!cache.snapshotsBySource) {
+                    return {}
+                }
+                return cache.snapshotsBySource
+            },
+        ],
+    })),
+    beforeUnmount(({ cache }) => {
+        // Clear the cache
+        if (cache.realTimePollingTimeoutID) {
+            clearTimeout(cache.realTimePollingTimeoutID)
+            cache.realTimePollingTimeoutID = undefined
+        }
+        cache.snapshotsBySource = undefined
+    }),
+])
